agent-project-main/pyproject.toml
[project]
name = "agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic",
    "python-dotenv",
]
